Mysite/requirements.txt
asgiref==3.2.7
astroid==2.4.0
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.4
django-allauth==0.41.0
django-crispy-forms==1.9.0
django-environ==0.4.5
django-filter==2.2.0
django-filters==0.2.1
djangorestframework==3.11.0
environ==1.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pylint==2.5.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
